+++ src/Constants.ts
@@ -9,7 +9,7 @@
 
 export const DEFAULT_SIYUAN_LANG = window.siyuan.config.lang
 export const isDev = process.env.DEV_MODE === "true"
-export const DEFAULT_SIYUAN_API_URL = "http://127.0.0.1:6806"
+export const DEFAULT_SIYUAN_API_URL = window.location.origin
 export const DEFAULT_SIYUAN_AUTH_TOKEN = ""
 export const DEFAULT_SIYUAN_COOKIE = ""
 export const SHARE_PRO_STORE_NAME = "share-pro.json"