+++ src/composables/useSiyuanApi.ts
@@ -19,15 +19,19 @@ import { ShareProConfig } from "../models/ShareProConfig"
 export const useSiyuanApi = (cfg: ShareProConfig) => {
   const logger = simpleLogger("use-siyuan-api", "share-pro", isDev)
 
+  if (cfg.siyuanConfig.apiUrl !== window.location.origin) {
+    logger.warn("siyuan api url not match, use default")
+    cfg.siyuanConfig.apiUrl = window.location.origin
+  }
   const siyuanConfig = new SiyuanConfig(cfg.siyuanConfig.apiUrl, cfg.siyuanConfig.token)
   // 开启了授权码可能不可用
   siyuanConfig.cookie = cfg.siyuanConfig.cookie
   // 一些常用设置
   siyuanConfig.preferenceConfig.fixTitle = cfg.siyuanConfig?.preferenceConfig?.fixTitle ?? false
-  siyuanConfig.preferenceConfig.docTreeEnable = cfg.siyuanConfig?.preferenceConfig?.docTreeEnable ?? true
-  siyuanConfig.preferenceConfig.docTreeLevel = cfg.siyuanConfig?.preferenceConfig?.docTreeLevel ?? 3
-  siyuanConfig.preferenceConfig.outlineEnable = cfg.siyuanConfig?.preferenceConfig?.outlineEnable ?? true
-  siyuanConfig.preferenceConfig.outlineLevel = cfg.siyuanConfig?.preferenceConfig?.outlineLevel ?? 6
+  siyuanConfig.preferenceConfig.docTreeEnable = cfg.appConfig?.docTreeEnabled ?? true
+  siyuanConfig.preferenceConfig.docTreeLevel = cfg.appConfig?.docTreeLevel ?? 3
+  siyuanConfig.preferenceConfig.outlineEnable = cfg.appConfig?.outlineEnabled ?? true
+  siyuanConfig.preferenceConfig.outlineLevel = cfg.appConfig?.outlineLevel ?? 6
   siyuanConfig.preferenceConfig.removeFirstH1 = true
   siyuanConfig.preferenceConfig.removeMdWidgetTag = true
   const blogApi = new SiYuanApiAdaptor(siyuanConfig)