+++ src/libs/pages/ShareManage.svelte
@@ -12,9 +12,12 @@
   import { onMount } from "svelte"
   import { ShareService } from "../../service/ShareService"
   import { simpleLogger } from "zhi-lib-base"
-  import { isDev, SHARE_LIST_PAGE_SIZE } from "../../Constants"
+  import { isDev, SHARE_LIST_PAGE_SIZE, SHARE_PRO_STORE_NAME } from "../../Constants"
   import Bench from "../components/bench/Bench.svelte"
   import { confirm, openTab, showMessage } from "siyuan"
+  import { ShareProConfig } from "../../models/ShareProConfig"
+  import { syncAppConfig } from "../../utils/ShareConfigUtils"
+  import { SettingService } from "../../service/SettingService"
 
   const logger = simpleLogger("share-manage", "share-pro", isDev)
   export let pluginInstance: ShareProPlugin
@@ -22,6 +25,7 @@
   const shareService = new ShareService(pluginInstance)
   let docs = []
   let loading = false
+  let settingConfig: ShareProConfig
 
   let tableData,
     tableLimit = SHARE_LIST_PAGE_SIZE,
@@ -31,13 +35,13 @@
     tableSearch
 
   const tableColumns = [
-    { id: "docId", name: pluginInstance.i18n.manage.columnDocId, hidden: false },
-    {
-      id: "author",
-      name: pluginInstance.i18n.manage.columnAuthor,
-      sort: false,
-      onClick: (e) => {},
-    },
+    // { id: "docId", name: pluginInstance.i18n.manage.columnDocId, hidden: false },
+    // {
+    //   id: "author",
+    //   name: pluginInstance.i18n.manage.columnAuthor,
+    //   sort: false,
+    //   onClick: (e) => {},
+    // },
     {
       id: "title",
       name: pluginInstance.i18n.manage.columnTitle,
@@ -107,14 +111,16 @@
       tableData = {
         results: docs.map((doc) => {
           return {
-            docId: doc.docId,
-            author: keyInfo.email,
+            // docId: doc.docId,
+            // author: keyInfo.email,
             title: doc.data.title,
             media_count: doc.media.length,
             createdAt: doc.createdAt,
             status: doc.status,
             action: `
             <a href="javascript:;" onclick="window.cancelShareFromSharePro('${doc.docId}','${doc.data.title}')">${pluginInstance.i18n.manage.actionCancel}</a>&nbsp;&nbsp;
+            <a href="javascript:;" style="${settingConfig?.appConfig?.homePageId===doc.docId?'color:green;text-decoration:none;cursor:text;':''}" onclick="window.setHomeFromSharePro('${doc.docId}','${doc.data.title}',${settingConfig?.appConfig?.homePageId===doc.docId})">${settingConfig?.appConfig?.homePageId===doc.docId?pluginInstance.i18n.manage.actionSetAlready:pluginInstance.i18n.manage.actionSetHome}</a>&nbsp;&nbsp;
+            <a href="javascript:;" onclick="window.viewDocFromSharePro('${doc.docId}','${doc.data.title}')">${pluginInstance.i18n.manage.actionViewDoc}</a>&nbsp;&nbsp;
             <a href="javascript:;" onclick="window.goToOriginalDocFromSharePro('${doc.docId}')">${pluginInstance.i18n.manage.actionGotoDoc}</a>
             `,
           }
@@ -130,7 +136,7 @@
   }
 
   // @ts-ignore
-  window.cancelShareFromSharePro = async function (docId: string, docTitle) {
+  window.cancelShareFromSharePro = async function (docId: string, docTitle: string) {
     confirm(pluginInstance.i18n.tipTitle, pluginInstance.i18n.confirmDelete + "【" + docTitle + "】", async () => {
       const ret = await shareService.deleteDoc(docId)
       if (ret.code === 0) {
@@ -140,6 +146,40 @@
         showMessage(pluginInstance.i18n.topbar.cancelError + ret.msg, 7000, "error")
       }
     })
+  }
+
+  // @ts-ignore
+  window.setHomeFromSharePro = async function (docId: string, docTitle: string, isSet: boolean) {
+    if(isSet){
+      return
+    }
+
+    const msgSetHomeConfirmWithParam = pluginInstance.i18n.manage.setHomeConfirm
+    const msgSetHomeConfirm = msgSetHomeConfirmWithParam.replace("[param1]", docTitle)
+    confirm(pluginInstance.i18n.tipTitle, msgSetHomeConfirm, async () => {
+      settingConfig.appConfig.homePageId = docId
+      await pluginInstance.saveData(SHARE_PRO_STORE_NAME, settingConfig)
+      try {
+        const settingService = new SettingService(pluginInstance)
+        await syncAppConfig(settingService, settingConfig)
+        await updateTable()
+        showMessage(`${pluginInstance.i18n.manage.setHomeSuccess}`, 2000, "info")
+      } catch (e) {
+        showMessage(`${pluginInstance.i18n.manage.setHomeError}${e}`, 7000, "error")
+      }
+    })
+  }
+
+  // @ts-ignore
+  window.viewDocFromSharePro = async function (docId: string, _docTitle: string) {
+    const docInfo = await shareService.getSharedDocInfo(docId)
+    const shareData = docInfo?.data ? JSON.parse(docInfo.data) : null
+    if (!shareData) {
+      const noShareMsg = pluginInstance.i18n.topbar.msgNoShare + docInfo.msg
+      showMessage(noShareMsg, 7000, "error")
+      return
+    }
+    window.open(shareData.viewUrl)
   }
 
   // @ts-ignore
@@ -153,6 +193,7 @@
   }
 
   onMount(async () => {
+    settingConfig = await pluginInstance.safeLoad<ShareProConfig>(SHARE_PRO_STORE_NAME)
     await updateTable()
   })
 
@@ -205,7 +246,9 @@
     flex-direction: column
     align-items: center
     justify-content: center
-    z-index: 1000 /* 确保蒙版层在最上层 */
+    z-index: 1000
+
+  /* 确保蒙版层在最上层 */
 
   .loading-indicator
     display: flex
@@ -219,7 +262,9 @@
     width: 20px
     height: 20px
     animation: spin 1s linear infinite
-    margin-right: 10px /* 使文字与加载图标之间有间距 */
+    margin-right: 10px
+
+  /* 使文字与加载图标之间有间距 */
 
   @keyframes spin
     0%