+++ src/libs/pages/ShareSetting.svelte
@@ -17,6 +17,7 @@
   import { onMount } from "svelte"
   import { KeyInfo } from "../../models/KeyInfo"
   import CustomSetting from "./setting/CustomSetting.svelte"
+  import DocSetting from "./setting/DocSetting.svelte"
 
   const logger = simpleLogger("share-main", "share-pro", isDev)
   export let pluginInstance: ShareProPlugin
@@ -51,6 +52,17 @@
       props: {
         pluginInstance: pluginInstance,
         dialog: dialog,
+        vipInfo: vipInfo,
+      },
+    })
+    // 文档设置
+    tabData.push({
+      label: pluginInstance.i18n.docSetting,
+      content: DocSetting,
+      props: {
+        pluginInstance: pluginInstance,
+        dialog: dialog,
+        vipInfo: vipInfo,
       },
     })