+++ src/models/AppConfig.ts
@@ -50,6 +50,16 @@ interface AppConfig {
     }
   ]
 
+  domains: Array<string>
+  domain: string
+  basePaths: Array<string>
+  docPath: string
+
+  docTreeEnabled: boolean
+  docTreeLevel: number
+  outlineEnabled: boolean
+  outlineLevel: number
+
   // 加上字符串索引签名，兼容 AppConfigInput 约束
   [key: string]: any
 }