+++ src/service/ShareService.ts
@@ -41,9 +41,29 @@ class ShareService {
     return await this.shareApi.getVipInfo(token)
   }
 
-  public async createShare(docId: string) {
+  public async createShare(docId: string, singleDocSetting?: any) {
     try {
-      const blogApi = await this.getSiyuanApi()
+      const cfg = await this.pluginInstance.safeLoad<ShareProConfig>(SHARE_PRO_STORE_NAME)
+      // ===============================================================================================================
+      // 上面是通用配置
+      const { kernelApi } = useSiyuanApi(cfg)
+      // 保存单篇文档的配置
+      if (singleDocSetting) {
+        const { docTreeEnable, docTreeLevel, outlineEnable, outlineLevel } = singleDocSetting
+        cfg.siyuanConfig.preferenceConfig = {
+          ...cfg.siyuanConfig.preferenceConfig,
+          docTreeEnable: docTreeEnable ?? cfg.siyuanConfig.preferenceConfig.docTreeEnable,
+          docTreeLevel: docTreeLevel ?? cfg.siyuanConfig.preferenceConfig.docTreeLevel,
+          outlineEnable: outlineEnable ?? cfg.siyuanConfig.preferenceConfig.outlineEnable,
+          outlineLevel: outlineLevel ?? cfg.siyuanConfig.preferenceConfig.outlineLevel,
+        }
+        await kernelApi.setSingleBlockAttr(docId, "share-pro-setting", JSON.stringify(singleDocSetting))
+      } else {
+        await kernelApi.setSingleBlockAttr(docId, "share-pro-setting", JSON.stringify({}))
+      }
+      // 在这里可以重写单篇文档
+      // ===============================================================================================================
+      const blogApi = await this.getSiyuanApi(cfg)
       const post = await blogApi.getPost(docId)
       this.logger.debug("get post", post)
       const sPost = new Post()
@@ -130,8 +150,7 @@ class ShareService {
   // private function
   // ================
 
-  private async getSiyuanApi() {
-    const cfg = await this.pluginInstance.safeLoad<ShareProConfig>(SHARE_PRO_STORE_NAME)
+  private async getSiyuanApi(cfg: ShareProConfig) {
     const { blogApi } = useSiyuanApi(cfg)
     return blogApi
   }