+++ CHANGELOG.md
@@ -1,7 +1,22 @@
 # Changelog
 
-## [1.6.0](https://github.com/terwerinc/siyuan-plugin-share-pro/compare/share-pro-v1.5.0...share-pro-v1.6.0) (2024-11-13)
+## [1.8.2](https://github.com/terwerinc/siyuan-plugin-share-pro/compare/share-pro-v1.8.1...share-pro-v1.8.2) (2024-11-29)
+### Bug Fixes
+* setting error in some cases ([e553342](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/e553342f9bf305f0112d6fdcc5aaba6abbbfd133))
+## [1.8.1](https://github.com/terwerinc/siyuan-plugin-share-pro/compare/share-pro-v1.8.0...share-pro-v1.8.1) (2024-11-25)
+## [1.8.0](https://github.com/terwerinc/siyuan-plugin-share-pro/compare/share-pro-v1.7.0...share-pro-v1.8.0) (2024-11-19)
 ### Features
+* split setting ([161d2b2](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/161d2b23c46bbeb80a04029d49227f6cc6d22767))
+* support dark mode ([6dbfb2b](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/6dbfb2b568a05ff8a73fefc95b49c3cdfe6e8a79))
+* support lang setting ([30f62de](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/30f62ded923d2d723a3a981e9b3e89ae6d7a522c))
+## [1.7.0](https://github.com/terwerinc/siyuan-plugin-share-pro/compare/share-pro-v1.6.0...share-pro-v1.7.0) (2024-11-18)
+* add custom share (experimental) ([a29ec59](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/a29ec59ad4220f2f1329403f56dbbbd7a0132df4))
+* fix siyuan api url error ([b3c0598](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/b3c0598e8c0488c6a293569e0affc4e5a5765e89))
+* support doc path config ([b874d1c](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/b874d1cb18553a8a73f9681a21048150160ede56))
+* support domain config ([8c6f252](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/8c6f2524bba1981a751213e119495ac0ebef9705))
+* support set home for custom domain users ([a047fbd](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/a047fbd648c959997ce02931a27aad6cf6c158e7))
+* support setting for doctree and outline ([24a2fd1](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/24a2fd149ab88a2d6ef1958beded9b8999bf7c2b))
+## [1.6.0](https://github.com/terwerinc/siyuan-plugin-share-pro/compare/share-pro-v1.5.0...share-pro-v1.6.0) (2024-11-13)
 * new ui/ux (experimental) - add tab component ([db1d189](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/db1d189a0341b7496306010da7ea0df44d75bb20))
 * support doctree and outline ([fb0444b](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/fb0444b9b98c6a0635f5d2d8a91e5c15136ec555))
 ### Miscellaneous
@@ -14,7 +29,6 @@
 * support pink-room theme ([7728120](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/77281207a6778db5aaf46f80e6c32b2025b11c46))
 * support tsundoku theme ([d19f343](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/d19f343b5d87385844ff6f6c7eb777d4a744f132))
 * update themes ([48f6d7d](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/48f6d7d4fb1b108746d4ae40803a4d156caf5e7e))
-### Bug Fixes
 * remove unused setting ([1bac8de](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/1bac8de36b4ec5ed74309fb96a9ccba0112f7c62))
 * zhihu theme code block ([fd898a0](https://github.com/terwerinc/siyuan-plugin-share-pro/commit/fd898a02a2423196d656e62a82d20dd07b62c731))
 ## [1.3.3](https://github.com/terwerinc/siyuan-plugin-share-pro/compare/share-pro-v1.3.2...share-pro-v1.3.3) (2024-11-05)